adding everything needed to display notes based on userId. created requiresAuth middleware to check that the user is authenticated, added userId field to note schema, implemented requiresAuth for all notes apis, and for getAuthenticatedUser route.

+++ backend/src/controllers/notes.ts
@@ -2,12 +2,15 @@ import type { RequestHandler } from "express";
 import NoteModel from "../models/note.ts";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined.ts";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+    
     try {
-        // throw Error("BOOOOM");
-        //throw createHttpError(401);
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+        
+        const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error); // ← This jumps to error middleware
@@ -16,8 +19,11 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -27,6 +33,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
             throw createHttpError(404, "Note not found");
         }
         
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "Not authorize to access this note");
+        }
+
         res.status(200).json(note);
     } catch (error) {
         next(error);
@@ -41,13 +51,17 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
-    
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!title) {
             throw createHttpError(400, "Note must have a title");
         }
 
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         });
@@ -71,7 +85,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -84,6 +102,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         if (!note){
             throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "Not authorize to access this note");
         }
 
         note.title = newTitle;
@@ -101,8 +123,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async(req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
@@ -113,6 +138,10 @@ export const deleteNote: RequestHandler = async(req, res, next) => {
             throw createHttpError(404, "Note not found");
         }
 
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "Not authorize to access this note");
+        }
+        
         await NoteModel.deleteOne({_id: noteId});
 
         res.sendStatus(204);
